파일구조 개선
